+++ src/Telegram.Bot/Types/InputFiles/InputMedia.cs
@@ -23,14 +23,24 @@ namespace Telegram.Bot.Types
         }
 
         /// <summary>
-        /// Initializes media with either a file_id or a HTTP url
+        /// Initializes an instance of <see cref="InputMedia"/> with either a file_id or a HTTP URL
         /// </summary>
-        /// <param name="value">file_id to send a file that exists on the Telegram servers or an HTTP URL for Telegram to get a file from the Internet</param>
+        /// <param name="value">
+        /// file_id to send a file that exists on the Telegram servers or an HTTP URL for Telegram to get a file
+        /// from the Internet
+        /// </param>
         public InputMedia(string value)
             : base(value)
         {
         }
 
+        /// <summary>
+        /// Initializes an instance of <see cref="InputMedia"/> with either a file_id or a HTTP URL
+        /// </summary>
+        /// <param name="value">
+        /// file_id to send a file that exists on the Telegram servers or an HTTP URL for Telegram to get a file
+        /// from the Internet
+        /// </param>
         public static implicit operator InputMedia(string value) =>
             value == null
                 ? null