+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Telegram Passport support
+- Type `EncryptedCredentials`
+- Type `EncryptedPassportElement`
+- Type `PassportData`
+- Type `PassportFile`
+- Property `ITelegramBotClient.BotId`
+- Property `Message.PassportData`
+
+### Changed
+
+- Marked `DownloadFileAsync(string,CancellationToken)` obsolete.
+
 ## [14.9.0] - 2018-08-06
 
 ### Added